% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{package}
\name{BGmiscinternal-package}
\alias{BGmiscinternal}
\alias{BGmiscinternal-package}
\title{BGmiscinternal: Internal extension of BGmisc}
\description{
This is an internal extension package for functions that are helpful to the research team. They either require too many dependencies for BGmisc or are very bespoke.
}
\author{
\strong{Maintainer}: S. Mason Garrison \email{garrissm@wfu.edu} (\href{https://orcid.org/0000-0002-4804-6003}{ORCID})

Authors:
\itemize{
  \item S. Mason Garrison \email{garrissm@wfu.edu} (\href{https://orcid.org/0000-0002-4804-6003}{ORCID})
}

}
\keyword{internal}
