% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not-In Operator}
\usage{
\%notin\% ...
}
\arguments{
\item{...}{A vector of values to be matched against (typically a constant or a column}
}
\value{
A logical vector the same length as \code{x}: \code{TRUE} where
  \code{x[i]} is not in \code{table}, \code{FALSE} otherwise.
}
\description{
A convenience operator that returns \code{TRUE} for each element of \code{x}
that is \emph{not} present in \code{table}.  It is the negation of \code{\%in\%}.
}
\keyword{internal}
