% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{package}
\name{BGmiscinternal-package}
\alias{BGmiscinternal}
\alias{BGmiscinternal-package}
\title{BGmiscinternal: Internal extension of BGmisc}
\description{
This is an internal extension package for functions that are helpful to the research team, but either require too many dependencies for BGmisc or are very bespoke
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

Authors:
\itemize{
  \item Jane Doe \email{jane@example.com}
}

}
\keyword{internal}
