% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlateOutcomesByGroup_helpers.R,
%   R/helpers.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not-In Operator}
\usage{
... \%notin\% NA

... \%notin\% NA
}
\arguments{
\item{x}{A vector of values to test.}

\item{table}{A vector of values to test against.}
}
\value{
A logical vector the same length as \code{x}.

A logical vector the same length as \code{x}: \code{TRUE} where
  \code{x[i]} is not in \code{table}, \code{FALSE} otherwise.
}
\description{
Returns \code{TRUE} for each element of \code{x} that is \emph{not} present
in \code{table}.  Equivalent to \code{!(\%in\%)}.

A convenience operator that returns \code{TRUE} for each element of \code{x}
that is \emph{not} present in \code{table}.  It is the negation of \code{\%in\%}.
}
\keyword{internal}
