% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World}
\usage{
hello()
}
\value{
Called for its side effect of printing to the console.
  Returns the printed string invisibly.
}
\description{
A simple example function that prints \code{"Hello, world!"} to the console.
}
\details{
You can learn more about package authoring with RStudio at:
\url{https://r-pkgs.org}
}
\examples{
hello()
}
